+++ src/app/globals.css
@@ -13,7 +13,7 @@
 body {
   margin: 0;
   padding: 0;
-  background-image: url('/images/cards-background.jpg'); /* Imagen de fondo */
+  background-image: url('/images/backgound.jpeg'); /* Imagen de fondo */
   background-size: cover;
   background-position: center;
   font-family: 'Montserrat', sans-serif;
@@ -114,10 +114,9 @@ nav a {
   width: 80%; /* Aumentar el ancho */
   background: linear-gradient(to left, #f7faf77a, #ffffff77);
   color: #4A4A4A;
-  border: 4px solid #ffffff; /* Cambiar el borde a 4px */
   display: flex; /* Cambiar a flexbox */
   justify-content: space-between; /* Espacio entre columnas */
-  padding: 16px;
+  padding: 50px;
   gap: 16px;
   border-radius: 10px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
@@ -321,6 +320,7 @@ footer .wave#wave4 {
   color: white;
   padding: 10px;
   cursor: pointer;
+
 }
 
 .carousel-indicators {
@@ -651,3 +651,58 @@ button:active {
   }
 }
 
+.cards-grid, .cards-grid-metodologia {
+  display: grid;
+  gap: 1rem;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  margin: 1rem 0;
+}
+
+.card-problema, .card-justificacion, .card-metodologia {
+  background-color: #f4f4f488;
+  padding: 1.2rem;
+  border-radius: 16px;
+  box-shadow: 0 4px 12px rgba(0,0,0,0.1);
+  transition: transform 0.3s ease;
+  text-align: center;
+}
+
+.card-problema:hover, .card-justificacion:hover, .card-metodologia:hover {
+  transform: translateY(-5px);
+  background-color: #e6f7ff8b;
+}
+
+.img-sensor {
+  width: 30%;
+  max-width: 600px;
+  margin: 1rem auto;
+  display: block;
+  border-radius: 12px;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.2);
+}
+
+
+
+.info{
+  color:white;
+  width: 80%;
+  text-align: center;
+  margin: auto;
+  text-justify: auto;
+}
+
+.imagenes-resultados img{
+  width: 10%;
+  object-fit: cover;
+  border-radius: 10px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+
+}
+
+.titulo-seccion{
+  font-size: 2.5em;
+  color: #fff;
+  text-align: center;
+  margin: 6%;
+  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5); /* Sombra del texto */
+}
